postgres-mcp-server/pyproject.toml
[tool.poetry]
name = "postgres-mcp-server"
version = "0.1.0"
description = "PostgreSQL MCP Server for database operations"
authors = ["Andres Vourakis"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
psycopg2-binary = "^2.9.10"
mcp = "^1.16.0"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
